Creating a simple plan of the budget App
Here is the Plan where i tryed to described an architecture of the budget App.
I will create 3 modules: Budget Controller, UI Controller and Global App Controller.
So, in the Budget Controller we will create object's prototypes both for incomes and expenses and the objects with arrays to fill up our income and expenses positions. 
Also there will be  functions to do calculation with numbers, adding a special ID number to each position (so that in later it will be possible to delete item and recalculate budget correctly)
There i will create functions to add items, calculate budget, calculate total budget and getBudget to put all total budget values, so that we can use it futher in the global app controller and display on the UI.
UI Controller will be made to interract with users, do some DOM manipulations, and here i will creat functions such as:
getInput (to use all the values that were inputed)
AddListItem (to add new html that later will be displayed items with our inputed values) 
clearFields (to clear all the fields when click button and get our values, so that then we won't do it be hands)
displayBudget (displayed fileds with our inputed values).
These 2 modules are closed for each other and will not interact with each other. 
So that i will create a 3 module called Controller to put them all together and, by creating Event Listeners (when we push the button on ENTER) do some code that we need to calculate and display results, combining these two modules (BudgetController and UIController). Also there will be another function updateBudget, to make some budget updates, by interracting two closed modules.
And last thing is init function, which I will create to set up my App.